Copy over config files in Trapdoor.prepare
Before, it was copied over in get_stats, which caused a little bit of a
problem in simulate_trapdoor_pr.py (shutil.copy would copy and paste the
same file).
Now, the config files are read in and copied in `prepare`, which is
executed before get_stats, preventing the problem.
Still not the nicest solution b/c trapdoor.cfg has to be read in (using
json) separately in `prepare`.